Refactors sensitivity module into uncertainty
