+++ kinetics/__init__.py
@@ -1,7 +1,7 @@
 from kinetics.Model import Model, Metrics, uM_to_mgml
 from kinetics.Genetic_Algorithm import GA_Base_Class
 from kinetics.Reaction_Classes import *
-from kinetics import Uncertainty, Sensitivity, Sampling
+from kinetics import Uncertainty, Sampling
 
 
 __version__ = '1.3.6'
